feat: Make flamegraph item a functional component
